feat: move listing of all topics to specific nav to reserve home page for specified project requirements
