+++ web/src/main.tsx
@@ -4,11 +4,13 @@ import App from './App.tsx';
 import { BrowserRouter } from 'react-router-dom';
 import { MantineProvider } from '@mantine/core';
 import { appTheme } from './constants/theme.tsx';
+import { Notifications } from '@mantine/notifications';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <BrowserRouter>
       <MantineProvider withGlobalStyles withNormalizeCSS theme={appTheme}>
+        <Notifications />
         <App />
       </MantineProvider>
     </BrowserRouter>
