+++ web/src/navigation/PrimaryNavigation.tsx
@@ -18,7 +18,7 @@ export function PrimaryNavigation(): React.ReactElement {
         </NavButton>
         <Flex align={'center'} gap={25}>
           <Flex gap={10}>
-            <NavButton route={routes.adminPortal}>Admin Portal</NavButton>
+            <NavButton route={routes.allTopics}>All Topics</NavButton>
           </Flex>
 
           <Menu>
